requirements.txt
# Core Libraries for naturaDock V1
# Note: Versions should be tested and locked before major releases.

rdkit-pypi==2022.9.5
biopython==1.81
pandas>=2.1.0
numpy==1.25.2
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
py3dmol==2.0.3
meeko==0.5.0
pdbfixer==1.11
openmm==8.1.1
toml==0.10.2



# Development & Code Quality
black==23.7.0
ruff==0.0.286
pytest==7.4.3